progress on extracting out services to serve up the behavior for the controllers
