+++ source/SIL.AppBuilder.Portal.Frontend/src/lib/auth/require-auth-helper.tsx
@@ -1,0 +1,14 @@
+import * as React from 'react';
+import { withRouter, RouterProps } from 'react-router';
+
+export function requireAuthHelper(authenticationChecker: (props: RouterProps) => any) {
+  return (props: any) => {
+    if (props.history) {
+      return authenticationChecker(props);
+    }
+
+    const WrappedWithRouter = withRouter(authenticationChecker);
+
+    return <WrappedWithRouter {...props } />;
+  };
+}
